+++ app/src/main/java/com/example/drivingefficiencyapp/MainMenu.kt
@@ -1,4 +1,6 @@
 package com.example.drivingefficiencyapp
+
+import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
@@ -9,14 +11,14 @@ class MainMenu : AppCompatActivity() {
         supportActionBar?.hide()
         setContentView(R.layout.main_menu)
 
-        //val startDriveButton = findViewById<Button>(R.id.startDriveButton)
+        val startDriveButton = findViewById<Button>(R.id.startDriveButton)
         //val viewTripsButton = findViewById<Button>(R.id.viewTripsButton)
 
-        // Starts driving activity, but doesn't exist yet
-        //startDriveButton.setOnClickListener {
-            // val intent = Intent(this, DriveActivity::class.java)
-            // startActivity(intent)
-        //}
+        //Starts driving activity, call to start drive file
+        startDriveButton.setOnClickListener {
+            val intent = Intent(this, StartDrive::class.java)
+            startActivity(intent)
+        }
 
         // Start trip activity, but doesn't exist yet
         //viewTripsButton.setOnClickListener {